Move runtime code into Runtime folder
