+++ Cql/Cql.Runtime/Cql.Runtime.csproj
@@ -4,7 +4,7 @@
 
 	<PropertyGroup>
 		<AssemblyName>Hl7.Cql.Runtime</AssemblyName>
-		<RootNamespace>Hl7.Cql.Runtime</RootNamespace>
+		<RootNamespace>Hl7.Cql</RootNamespace>
 		<Product>CQL engine runtime</Product>
 		<Description>Support package for Hl7.Cql. Contains the required runtime components for CQL execution.</Description>
 		<Configurations>Debug;Release</Configurations>
