+++ Cql/Cql.Runtime/Runtime/ErroredEnumerationContinuation.cs
