+++ Cql/Cql.Runtime/Runtime/ExceptionHandlingMethods.cs
