+++ Cql/Cql.Runtime/Runtime/Hosting/ServiceCollectionExtensions.cs
