+++ Cql/Cql.Runtime/Runtime/Logging/LoggingServiceCollectionExtensions.cs
