+++ Cql/Cql.Runtime/Runtime/StringExtensions.cs
